Отображение электронных фотографий в поле выбора в админке

+++ backend/apps/cart/api/v1/serializer.py
@@ -64,7 +64,7 @@ class CartPhotoLineCreateUpdateSerializer(serializers.Serializer):
         quantity = 0
         for photo_in_cart in photos_in_cart:
             quantity += photo_in_cart.get('quantity')
-        if quantity == 0:
+        if quantity == 0 and not validated_data["is_digital"] and not validated_data["is_photobook"]:
             validated_data.pop('cart').delete()
             return CartPhotoLine()
